Artmonitors-Autoupload/requirements.txt
# definitely important
Pillow==5.2.0
certifi==2018.8.24
chardet==3.0.4
pycryptodome==3.7.0
requests==2.20.0
urllib3==1.23

# probably important but I don't know exactly they're for
idna==2.7
setuptools==39.1.0
